Add side bar, content header, and form content
Added side bar using flexbox and added background image to it. Added
footer to provide image credit. Structured the content of the form
and styled the form inputs a bit so they had the proper layout. Having
issues with inputs not shrinking when browser is shrunk, but will
fix that in future update. Blue background on header of form content
was used to determine sizing for the section of the page based on the
viewport size.