Correct page and input sizing and grow/shrink
Corrected page sizing by setting html and body to 100% height and
width, while setting margins to 0px. Corrected the input sizing to
be 25% of stretched form div. Stretched the form div to ensure that
it fills the horizontal width. Background colors added for form div and
the labels/inputs to verify spacing and will be removed with later
update to ensure the form looks the way I want later. Next update will
include Odin logo and the bottom of the form.