Add logo and polish page design
Added logo and transparent background to side bar area. Still need to
size the div properly to ensure background isn't too large behind the
logo and text. Still need to fix bottom div beneath the form to fill
the rest of the page as it still does not want to grow to fill the open
space.