Polish logo styling
Corrected an issue where the logo and text would get covered up by the
content of the page by added a min-height and min-width to the logo
image. This will prevent the logo from being crushed at small sizes
and allowing the content to overlap the text of the logo div.