Add password match confirmation and styling
Added js function to check that passwords match and set the invalid
status to the form inputs. Tested with a couple of different inputs to
verify function. Couldn't figure out how to space the alert from the
form input itself, but it's a minor style issue that won't impact the
overall webpage so I didn't pursue it too intensly. Finished up the
project!